# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi Negocio Pro API",
        "description": "API para gestión completa de negocio: usuarios, productos, ventas, compras y más",
        "contact": {
            "email": "support@minegocio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server Local"
        }
    ],
    "paths": {
...
        "/api/sale-items": {
            "get": {
                "tags": ["Sale Items"],
                "summary": "Listar items de venta",
                "description": "Obtiene la lista de items de venta con filtros opcionales",
                "operationId": "getSaleItems",
                "parameters": [
                    {
                        "name": "sale_id",
                        "in": "query",
                        "description": "Filtrar por ID de venta",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filtrar por ID de producto",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de items de venta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": ["Sale Items"],
                "summary": "Crear nuevo item de venta",
                "description": "Crea un nuevo item de venta y actualiza el stock del producto",
                "operationId": "createSaleItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sale_id",
                                    "product_id",
                                    "quantity",
                                    "unit_price"
                                ],
                                "properties": {
                                    "sale_id": {
                                        "description": "ID de la venta",
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "description": "ID del producto",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Cantidad a vender",
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "unit_price": {
                                        "description": "Precio unitario",
                                        "type": "number",
                                        "format": "decimal",
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item de venta creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item de venta creado exitosamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o stock insuficiente"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido"
                    },
                    "404": {
                        "description": "Venta o producto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/sale-items/by-sale/{saleId}": {
            "get": {
                "tags": ["Sale Items"],
                "summary": "Obtener items de una venta específica",
                "description": "Obtiene todos los items que pertenecen a una venta específica",
                "operationId": "getSaleItemsBySale",
                "parameters": [
                    {
                        "name": "saleId",
                        "in": "path",
                        "description": "ID de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items de la venta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "summary": {
                                            "properties": {
                                                "total_items": {
                                                    "type": "integer"
                                                },
                                                "total_quantity": {
                                                    "type": "integer"
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "decimal"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido"
                    },
                    "404": {
                        "description": "Venta no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}